# frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Shift Ease API",
    "version": "v1"
  },
  "paths": {
...
    "/api/schedules/confirmed/my-shifts": {
      "get": {
        "tags": [
          "ScheduleEmployee"
        ],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedules/schedule-summaries": {
      "get": {
        "tags": [
          "ScheduleEmployer"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedules/schedule-management": {
      "get": {
        "tags": [
          "ScheduleEmployer"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedules/schedule": {
      "get": {
        "tags": [
          "ScheduleEmployer"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedules/confirmed": {
      "get": {
        "tags": [
          "ScheduleEmployer"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedules/create-empty-schedule": {
      "post": {
        "tags": [
          "ScheduleEmployer"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedules/update-schedule": {
      "post": {
        "tags": [
          "ScheduleEmployer"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BllSchedulePost"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BllSchedulePost"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BllSchedulePost"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedules/{scheduleId}/unconfirm": {
      "post": {
        "tags": [
          "ScheduleEmployer"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedules/{id}/export": {
      "get": {
        "tags": [
          "ScheduleEmployer"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedule-generator/generate-greedy": {
      "post": {
        "tags": [
          "ScheduleGenerator"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BllScheduleGenerateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BllScheduleGenerateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BllScheduleGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedule-generator/generate-aco": {
      "post": {
        "tags": [
          "ScheduleGenerator"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BllAcoScheduleGenerateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BllAcoScheduleGenerateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BllAcoScheduleGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedule-generator/generate-ga": {
      "post": {
        "tags": [
          "ScheduleGenerator"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BllGaScheduleGenerateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BllGaScheduleGenerateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BllGaScheduleGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/organizations/{orgId}/schedule-generator/generate-aco-ga": {
      "post": {
        "tags": [
          "ScheduleGenerator"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BllAcoGaScheduleGenerateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BllAcoGaScheduleGenerateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BllAcoGaScheduleGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BllAcoGaScheduleGenerateRequest": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "totalHours": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "hardTotalHours": {
            "type": "boolean"
          },
          "numAnts": {
            "type": "integer",
            "format": "int32"
          },
          "numAcoIterations": {
            "type": "integer",
            "format": "int32"
          },
          "populationSize": {
            "type": "integer",
            "format": "int32"
          },
          "numGaGenerations": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BllAcoScheduleGenerateRequest": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "totalHours": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "hardTotalHours": {
            "type": "boolean"
          },
          "numAnts": {
            "type": "integer",
            "format": "int32"
          },
          "numIterations": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "BllGaScheduleGenerateRequest": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "totalHours": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "hardTotalHours": {
            "type": "boolean"
          },
          "populationSize": {
            "type": "integer",
            "format": "int32"
          },
          "numGenerations": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "BllScheduleGenerateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "totalHours": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "hardTotalHours": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "BllSchedulePost": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "isConfirmed": {
            "type": "boolean"
          },
          "shifts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShiftPost"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EmployeeInShiftDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ShiftPost": {
        "type": "object",
        "properties": {
          "shiftTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "employees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeInShiftDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}